new files

+++ print_int.c
@@ -1,13 +1,13 @@
 #include "main.h"
 
 /**
- * find_int - function that finds string lenght of interger.
+ * find_int - a function that finds string lenght of interger.
  * @num: number to be printed.
  * @len: lenght of a string.
  * Return: return string length.
  */
 
-void find_int(int num, int *len)
+void find_int(long int num, int *len)
 {
 	if (num < 0)
 	{
@@ -22,7 +22,7 @@ void find_int(int num, int *len)
 }
 
 /**
- * print_int - function that prints argument pointers.
+ * print_int - a function that prints argument pointers.
  * @ap: argument pointer.
  * Return: return string length.
  */
@@ -30,7 +30,7 @@ void find_int(int num, int *len)
 int print_int(va_list ap)
 {
 	int len = 0;
-	int num = va_arg(ap, int);
+	long int num = (long int)va_arg(ap, int);
 
 	find_int(num, &len);
 	return (len);
